Crate db

Rust implmentation of MongoDB Client. One must have already started a MongoDB server, otherwise when attempting to create a handler to the database, it will panic.

Basic Usage

let mut mongo = DatabaseHandler::new("Caller".to_string()).unwrap(); // will panic if no mongodb server is running.
mongo.set_db("test").unwrap(); // sets internal database

Modules

templates
Defines the different structures that are serializable to MongoDB All structures that will have a place in MongoDB need to be defined here.

utils
List of utilities and common data structures

Structs

DatabaseHandler
Wrapper to MongoDB running in background