Module db::cache_structures

Collection of utilies that define structures to store in caching server as well as parsing those objects for server consumption

For every wrapper around a rust object that you would like to serialize into the caching server needs to have at least two fields to make it possible

redis::Connection

UID

Defining this trait on a struct will allow methods to become available where you can serialize/deserialize your rust object into redis server.Connection

The naming scheme is pretty simple, as it uses the following format.

MyStruct:UID:fieldName fieldValue

this simple use of redis get/set allows individual fields to contain meta data whereby hmap will not allow, such as expirary times for fields etc..

Modules

socket
Holds structure for Socket information and manipulation in caching server.

Traits

Cachable
trait to all structures that are cachable to cache server