Struct db::cache_structures::socket::CacheSocket

pub struct CacheSocket {
    pub uid: UID,
    pub ip: String,
    pub port: u16,
    // some fields omitted
}

Template structure to for raw socket information

Fields
uid: UID

unique id of socket

ip: String

ip address of where socket is coming form

port: u16

port of where socket is coming from

Methods

impl<'a> CacheSocket

pub fn new_with_uid(uid: UID) -> Self

create instance of socket with supplied UID

pub fn new() -> Self

create blank instance of Socket

pub fn set_address(&mut self, addr: &SocketAddr) -> &mut Self

set ip and port for this struct with a supplied std::net::SocketAddr

pub fn set_ip(&mut self, ip: &'a str) -> &mut Self

set ip for this struct

pub fn set_port(&mut self, port: u16) -> &mut Self

set port for this struct

pub fn set_value<F: ToRedisArgs + ToString, V: ToRedisArgs + ToString>(
&mut self,
field: F,
value: V
) -> RedisResult<()>

set key and value

pub fn get_value<T: FromStr + Debug>(
&mut self,
field_name: &'a str
) -> Option<T> where
T::Err: Debug,

retrieve a value from field of struct, if it does not exist, will return None