Struct db::clients::ClientHandler

pub struct ClientHandler;

Methods

impl ClientHandler

pub fn upsert(
&self,
conn: &PgConnection,
new_client: &Client
) -> QueryResult<Client>

Attempts to insert a new client with UID, if there is a conflic, it will update the record.

pub fn upsert_batch(
&self,
conn: &PgConnection,
clients: ClientVector
) -> QueryResult<()>

Attempts to insert a new client with UID, if there is a conflic, it will update the record. Doesn't work quite as expected.. Followed the guides from [here] (https://stackoverflow.com/questions/47626047/execute-an-insert-or-update-using-diesel) but doesn't seem to actually set the excluded value where the conflict happened..

pub fn remove_uid(&self, conn: &PgConnection, id: UID) -> QueryResult<usize>

Permanently remove record from table, by UID

pub fn remove_uids(
&self,
conn: &PgConnection,
uids: Vec<UID>
) -> QueryResult<usize>

Remove a list of UIDS from db by suppling a vec of UIDs careful if supplied vector is empty, it will remove all records in table

pub fn get_uid(&self, conn: &PgConnection, id: UID) -> QueryResult<Vec<Client>>

Get single UID from db

pub fn get_uids(
&self,
conn: &PgConnection,
uids: Vec<UID>
) -> QueryResult<Vec<Client>>

Retrieve a list of UIDS from db by suppling a vec of UIDs if an empty set of uids is supplied, it will retrieve all records in table

pub fn get_uids_range(
&self,
conn: &PgConnection,
from: UID,
to: UID
) -> QueryResult<Vec<Client>>

Get a range of UIDs Note that UIDS that do not exists will be logged, but will not show up in the returned vector

pub fn get_ip_exact<'a>(
&self,
conn: &PgConnection,
ip: &'a str
) -> QueryResult<Vec<Client>>

retrieve records with exact IP address

pub fn get_ip_like<'a>(
&self,
conn: &PgConnection,
pattern_match: String
) -> QueryResult<Vec<Client>>

retrieve a vector of ip address with the appropriate match pattern provided

let pattern = format!("%{}%", "168.0");

// vector of Clients objects with ips matching `*168.0*`
let results = ClientHandler::get_ip_like(&conn, pattern);