Struct db::DatabaseHandler

pub struct DatabaseHandler {
    pub client: Option<Client>,
    pub db: Option<Database>,
}

Wrapper to MongoDB running in background

Fields
client: Option<Client>

Holds the actual client handler to MongoDB

db: Option<Database>

Database handler within MongoDB set using DatabaseHandler::set_db()

Methods

impl DatabaseHandler

pub fn new(name: String) -> MongoResult<Self>

Create a new handler to MongoDB

pub fn db_name(&self) -> &str

retrieve current database name

pub fn set_db<'a>(&mut self, db: &'a str) -> MongoResult<()>

set database within client, all actions will be targeted to supplied database name.

pub fn get_db<'a>(&mut self) -> MongoResult<&Database>

Returns reference to Database Handler, if db is not set it will return Error;

pub fn get_collection<'a>(
&mut self,
collection: &'a str
) -> MongoResult<Collection>

Returns Collection within current database

pub fn collection_exists(&mut self, col: &Collection) -> MongoResult<bool>

Checks if collection exists within database

pub fn insert_one<T: Serialize + MongoDocument>(
&mut self,
object: &T,
collection: &Collection,
options: impl Into<Option<InsertOneOptions>>
) -> MongoResult<()>

Attempts to serialize a rust object into a valid mongo document. Object must have MongoDocument + Serialize traits in order to be considered valid

Will error if attempting to insert to a collection that doesn't exist on database

Basic Usage

type UID = u64;
struct Crab{
    uid: UID,
    name: String,
}

impl MongoDocument for Crab{
    fn name(&self) -> String{
        self.name.clone()
    }
     
    fn uid(&self) -> UID{
        self.uid
    }
}

let crab = Crab{
    uid: db::utils::gen_uid(),
    name: "Crab".to_string();
}

let mut mongo = DatabaseHandler::new("Caller".to_string()).unwrap();

mongo.set_db("entities").unwrap();

let collection = mongo.get_collection("mobs").expect("Couldn't find collection");

// serialized and inserted to MongoDB with Schema
// entities -> mobs -> crab_bson
mongo.insert_one(&crab, &collection, None).unwrap();

pub fn doc_exists<T: Serialize + MongoDocument>(
&mut self,
object: &T,
collection: &Collection,
filter_on: FilterOn,
options: impl Into<Option<CountOptions>>
) -> MongoResult<bool>

Check to see if a particular document exists within data base and supplied collection. Filters based on either UID which will always be unique, or name which could have multiple instances in collection.

let crab = Crab::new(); // valid mongo document serializable struct

let col = mongo.get_collection("mobs").unwrap();

assert_eq!(true, doc_exists(&crab, &col, FilterOn::UID, None))

pub fn get_docs<T: Serialize + MongoDocument>(
&mut self,
object: &T,
collection: &Collection,
filter_on: FilterOn,
find_options: impl Into<Option<FindOptions>>
) -> MongoResult<Cursor>

Attempts to get a Database Cursor to all documents specified by Object T

let cursor = get_docs(&crab, &mobs_collection, FilterOn::NAME, None).unwrap();

for result in cursor{
    assert_eq!(bson::Document, result);
}

pub fn get_doc<T: Serialize + MongoDocument>(
&mut self,
object: &T,
collection: &Collection,
filter_on: FilterOn,
find_options: impl Into<Option<FindOneOptions>>
) -> MongoResult<Option<Document>>

Attempts to return a single Document within Collection

let document = get_doc(&crab, &mob_collection, None).unwrap();

assert_eq!(bson::Document, document);

pub fn update<T: Serialize + MongoDocument>(
&mut self,
object: &T,
collection: &Collection,
filter_on: FilterOn,
find_options: impl Into<Option<FindOptions>>,
update_options: Option<UpdateOptions>
) -> MongoResult<()>

Updates documents in collection based on FilterOn Updates ALL instances based on FilterOn and FindOptions

let crab = Crab{
    uid: db::utils::gen_uid(),
    name: "little crab",
}

let crab_doc = mongo.get_doc(&crab, &mob_collection, None, None)?;
assert_neq!(crab, db::utils::bson_to_object(crab_doc)?);

mongo.update(
    &crab,
    &mob_collection,
    db::utils::FilterOn::UID,
    None,
    None,
)

let crab_doc = mongo.get_doc(&crab, &mob_collection, None, None)?;
assert_eq!(crab, bson_to_object(crab_doc)?);