Trait db::BlockProvider

pub trait BlockProvider: BlockHeaderProvider {
    fn block_number(&self, hash: &H256) -> Option<u32>;

    fn block_hash(&self, number: u32) -> Option<H256>;

    fn block(&self, block_ref: BlockRef) -> Option<Block>;

    fn block_transaction_hashes(&self, block_ref: BlockRef) -> Vec<H256>;

    fn block_transactions(&self, block_ref: BlockRef) -> Vec<Transaction>;

    fn contains_block(&self, block_ref: BlockRef) -> bool { ... }
}

Required Methods

fn block_number(&self, hash: &H256) -> Option<u32>

resolves number by block hash

fn block_hash(&self, number: u32) -> Option<H256>

resolves hash by block number

fn block(&self, block_ref: BlockRef) -> Option<Block>

resolves deserialized block body by block reference (number/hash)

fn block_transaction_hashes(&self, block_ref: BlockRef) -> Vec<H256>

resolves list of block transactions by block reference (number/hash)

fn block_transactions(&self, block_ref: BlockRef) -> Vec<Transaction>

returns all transactions in the block by block reference (number/hash)

Provided Methods

fn contains_block(&self, block_ref: BlockRef) -> bool

returns true if store contains given block