Trait db::TransactionOutputProvider

pub trait TransactionOutputProvider: Send + Sync {
    fn transaction_output(&self,
                          outpoint: &OutPoint,
                          transaction_index: usize)
                          -> Option<TransactionOutput>;
    fn is_spent(&self, outpoint: &OutPoint) -> bool;
}

Should be used to get canon chain transaction outputs.

Required Methods

fn transaction_output(&self,
outpoint: &OutPoint,
transaction_index: usize)
-> Option<TransactionOutput>

Returns transaction output.

fn is_spent(&self, outpoint: &OutPoint) -> bool

Returns true if we know that output is double spent.