Crate db

Modules

bytes
Wrapper around Vec<u8>

hash
Fixed-size hashes

kv

Structs

BestBlock
Best block information

BlockAncestors

BlockChainDatabase

BlockIterator

ForkChainDatabase

SideChainOrigin

TransactionMeta
structure for indexing transaction info

Enums

BlockOrigin

BlockRef

Error

Traits

AsSubstore
Allows casting Arc to reference to any substore type

BlockChain

BlockHeaderProvider

BlockProvider

CanonStore

ConfigStore
Configuration storage interface

ForkChain

Forkable

IndexedBlockProvider

Store
Blockchain storage interface

TransactionMetaProvider
Transaction meta provider stores transaction meta information

TransactionOutputProvider
Should be used to get canon chain transaction outputs.

TransactionProvider
Should be used to obtain all transactions from canon chain and forks.

Type Definitions

SharedStore